+++ frontend/package.json
@@ -46,6 +46,8 @@
     ]
   },
   "devDependencies": {
+    "@babel/plugin-syntax-class-properties": "^7.7.4",
+    "@babel/plugins/plugin-proposal-class-properties": "^7.7.4",
     "@babel/preset-env": "^7.7.4",
     "@babel/preset-react": "^7.7.4",
     "babel-jest": "^24.9.0",
