--- package.json
@@ -1,43 +1,0 @@
-{
-  "name": "simmonsdatavis",
-  "version": "1.0.0",
-  "description": "This web application is a survey taking and data visualization website where students complete a survey and can view their qualities, such as Creativity and Communication, displayed in a spider chart. Students can compare their results with other survey takers by adding dataset(s) by gender, race/ethnicity, and/or major. Students can also look at a bar chart which uses other students survey results on the activities that can improve specific qualities.",
-  "main": "server.js",
-  "directories": {
-    "test": "tests"
-  },
-  "dependencies": {
-    "concurrently": "^5.0.0",
-    "bcryptjs": "^2.4.3",
-    "body-parser": "^1.19.0",
-    "chai": "^4.2.0", 
-    "cors": "^2.8.5",
-    "express": "^4.17.1",
-    "is-empty": "^1.2.0",
-    "jsonwebtoken": "^8.5.1",
-    "mongoose": "^5.7.7",
-    "morgan": "^1.9.1",
-    "nodemailer": "^6.3.1",
-    "passport": "^0.4.0",
-    "passport-jwt": "^4.0.0",
-    "should": "^13.2.3",
-    "supertest": "^4.0.2",
-    "validator": "^12.0.0"
-  },
-  "devDependencies": {},
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js",
-    "heroku-postbuild": "cd frontend && npm install && npm run build"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/introNG2E/simmonsDataVis.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/introNG2E/simmonsDataVis/issues"
-  },
-  "homepage": "https://github.com/introNG2E/simmonsDataVis#readme"
-}
